minitalk finished

+++ 07 - minitalk/serveur.c
@@ -6,7 +6,7 @@
 /*   By: thgaugai <thgaugai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/06 09:20:05 by thgaugai          #+#    #+#             */
-/*   Updated: 2025/01/08 17:33:44 by thgaugai         ###   ########.fr       */
+/*   Updated: 2025/01/09 09:01:33 by thgaugai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ int	main(void)
 
 	pid_server = getpid();
 	ft_putstr_fd("Server pid : ", 1);
-	ft_putnbr_fd((int)pid_server, 1);
+	ft_putnbr_fd(pid_server, 1);
 	ft_putchar_fd('\n', 1);
 	sa.sa_handler = handle_signint;
 	sa.sa_flags = 0;
